+++ CHANGELOG.md
@@ -56,6 +56,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+- 2026-07-11 | 🐛 fix(mobile): localize weekly order screens
 - 2026-07-11 | 🐛 fix(bylaws): refine query experience
 - 2026-07-10 | 🐛 fix(android-ui): space screen header content
 - 2026-07-10 | 🐛 fix(community): merge profile intro copy